Switched to using long sequenceNumbers while still sending 2 bytes on the wire but then extending them back to 8 bytes on receipt.  This should take care of the duration of the connection without rollover

+++ components/gnutella-core/src/main/java/com/limegroup/gnutella/udpconnect/DataMessage.java
@@ -11,7 +11,7 @@ public class DataMessage extends UDPConnectionMessage {
     /**
      * Construct a new DataMessage with the specified data.
      */
-    public DataMessage(byte connectionID, int sequenceNumber, 
+    public DataMessage(byte connectionID, long sequenceNumber, 
 	  byte[] data, int datalength) 
       throws BadPacketException {
